Boilerplate for very simple game.
This is very basic and should be adjusted if Evgeni sees any necessary changes.

I have a class SimpleGame where I have a reset method that is also used to initilaize, a get_state method that fully reveals ALL information (including player 1 and player 2's card, which we maybe should change), a step method that takes in a bet amount to experiment with and takes a bet and fold action, and a method to determine the winner based on card values.